fixed linux build errors
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@5964 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/uFldWrapDetect/FldWrapDetect.cpp
@@ -96,7 +96,7 @@ bool FldWrapDetect::OnNewMail(MOOSMSG_LIST &NewMail)
 	m_xysegments[node_name].push_front(new_segment); 
 	
 	// trim segments down to allowed size
-	while(m_xysegments[node_name].size() > m_max_num_segments){
+	while(m_xysegments[node_name].size() > (unsigned int) m_max_num_segments){
 	  m_xysegments[node_name].pop_back();
 	} // end while
